INVISODEV-954 Change zoom to respect MAX_ZOOM_LEVEL (#2)
Even though `MAX_ZOOM_LEVEL` is set to 100 it is possible to end up with a zoom greater than 100 % as the scaleFactor provided to scale3d in some cases might be greater than 1. A simple check is added to make sure this property never gets larger than what's configured as the maximum zoom level.

I also noticed that even with a maximum zoom the scaleFactor would still not be 1 (100 %). In some cases it was like 0.95 instead. An extra condition is added to force the scaleFactor to be max if the zoom level is max.